requirements.txt
httpx
python-dotenv
brotli
fastapi
uvicorn
